Remove old parser code
