+++ docs/source/conf.py
@@ -26,6 +26,7 @@ extensions = [
     'aiida.sphinxext',
     'sphinxcontrib.apidoc',
     'sphinxext.remoteliteralinclude',
+    'myst_parser',
 ]
 
 # Add any paths that contain templates here, relative to this directory.
