+++ src/aiida_vasp/parsers/old/settings.py
@@ -11,10 +11,9 @@ from copy import deepcopy
 from dataclasses import dataclass
 from typing import List
 
-from pydantic import BaseModel
-
 from aiida_vasp.parsers.content_parsers import *
 from aiida_vasp.utils.extended_dicts import update_nested_dict
+from pydantic import BaseModel
 
 
 @dataclass
