+++ src/aiida_vasp/parsers/old/vasp2w90.py
@@ -7,7 +7,6 @@ AiiDA Parser for aiida_vasp.Vasp2w90Calculation.
 
 from aiida.orm.nodes.data.list import List
 from aiida.plugins import DataFactory
-
 from aiida_vasp.parsers.content_parsers.win import WinParser
 
 from .vasp import VaspParser
